Controllers: Refactor ControllerTree.h classes into new files
